Reduce the number of required project (re-)builds
Up until now, target projects were rebuilt after every JPF
instrumentation step. This was necessary because the instrumented
code needed to be built before the following JPF execution step.

To avoid having to rebuild the projects after every instrumentation
step, instrumentation steps for all tests are now performed as a
batch. After this instrumentation batch, a single build is enough to
prepare the project for the following JPF execution steps, which are
now executed as a batch as well.