Reduce (error) coupling across processing flows
Up until now, all further processing steps of a project were skipped
if any exception occurred for any of the project's processing tasks.

With this change, processing flows that are not directly affected
by an error can still continue even if other processing flows are
not successful.

For example, tests for which JPF execution is successful are now
generalized as usual even if there are other tests for which JPF
execution failed.