Improve handling of task dependencies in scheduling logic
Up until now, it was necessary to "manually" track which tasks were
successful or not to ensure appropriate scheduling of corresponding
follow-up tasks.

This lead to quite repetitive code because such dependencies need to
be checked many times throughout the overall program execution.

With this change, tasks themselves are now responsible for scheduling
of required follow-up tasks. This automatically ensures that follow-up
tasks are  only scheduled if their dependencies execute successfully.