Fix detection of tests covering JDK classes
First step towards supporting generalization for tests covering JDK
classes (#22).

Calling `toAst` on the ResolvedMethodDeclaration of a (native) JDK
class / method returns an empty Optional, so we need to get the
information that we are interested in directly from the
ResolvedMethodDeclaration without going the `toAst` route.

This works for everything except for the storage location of the
corresponding class, which is simply not available if we don't have
AST information => set the path to null for affected classes.