Add cleanup of old project information before a new run
During development, behavior across runs can change quite a bit as
code is modified, so it's quite convenient to have only the
information that the newest run produced.

Later down the line, it might be more appropriate to add a
configuration option that allows the user of the tool to manually
de-/activate the cleanup step as needed.

Closes #21