Move JavaParser initialization to ProjectSetupTask
When creating the CombinedTypeSolver for JavaParser, we assume
that source code and test code are stored in the src/main/java and
src/test/java directories of the target project.

If at least one of these directories does not exist, JavaParser
throws an exception such as "java.lang.IllegalStateException: SrcDir
does not exist or is not a directory".

Up until now, this exception was only output to the console. By
moving the JavaParser initialization into a task, any exceptions
that are thrown by it are now actually logged as well as stored in
the DB as part of the ProjectSetupTask task record.