Fix Exceptions caused by non-identifiable tested methods
There were exceptions if:

(i) the identified tested method could not be resolved (i.e., is not
part of the analyzed project but rather from some dependency)
(ii) no tested method could be identified at all.

These cases are not check for so there are no uncaught exceptions
anymore for them. However, no data can be collected then anyway,
so the corresponding database columns now need to be nullable.