Fix too early checking of main/testCompiledPath
The presence of the main/testCompiledPath was checked in the
ProjectSetupTask. However, when setting up the project, no main/test-
CompiledPath exists yet, because those are created by the the Project-
BuildTask (which comes AFTER the ProjectSetupTask).

To fix this, the checks were moved from the ProjectSetupTask to the
ProjectBuildTask.