Fix incompilable code for @Test(timeout = ...) annotations
Up until now, @Test annotations were simply renamed to @Property.

This caused incompilable code to be produced when the @Test annotation
had any (timeout) parameters.

To avoid this, we do not rename the existing annotation anymore but
instead remove the old annotation and add a completely new one. Any
parameters that were set on the old annotation are currently lost.