Fix params of called methods being included in in-/output specifications
Because the parameters of the tested method are the only ones that
we can directly control, they should also be the only ones that
are included in the input / output specifications.

The problem was that we marked the whole class as symbolic via the
JPF configuration (leading to params of other tested class methods
being included in the specifications), when we should have only
marked the tested method as symbolic.

The original idea behind marking the whole class as symbolic was to
collect constraints for not only the tested method, but also for any
"setup" methods that are called beforehand. However, the current
generalization logic is far too limited to support such cases.