Remove EvoSuite annotations etc. from generalized classes
The EvoSuite annotations force the annotated test class to be run
with the EvoRunner test execution engine.

However, the EvoRunner does not support the @Property annotations
used by jqwik, causing test execution attempts of the generalized
classes to fail with "No runnable methods" errors.

To ensure the generalized classes can be executed, we now simply
remove all EvoSuite annotations in the generalized classes.

While this means that we do not benefit from EvoSuite's improved test
isolation and reproducibility features in the generalized tests, it's
still a much more acceptable solution than not being able to execute
the generalized tests at all.