Improve robustness of build.gradle modifications
Until now, some modifications of the build.gradle file did not
have their own checks to determine whether they are necessary or not.

For example, the pitest _plugin_ was always added when the pitest
_dependency_ was found to be missing. Similarly, useJUnitPlatform()
was added to all JUnit 4 projects, even if it was already present
in a project's build.gradle file.

With this commit, several such cases now have their own checks
to avoid more such unnecessary or potentially build-breaking
build.gradle file modifications.

Logging for both Gradle and Maven projetcs was adapted accordingly
to report not only dependency changes but also all other changes
such as the ones listed above.