Copy build.gradle / pom.xml before modifying them
To make it easier to revert the generalization, the build file of the
target project (build.gradle for Gradle projects, pom.xml for Maven
projects) is now copied before modifying it.

All generalizations steps only modify and use the COPIED file. That
way, the original project build still works as normal for as long as
possible, i.e., until some code is added to the project that requires
dependencies which are missing in the original build file.