Preserve collected data during cleanup
The CleanupTask now only removes created test and build files, but
preserves all data that was collected during the processing and
stored in the DB and the data directory.

This ensures that we can compare generalization results (runtime,
coverage, mutation score, ...) across multiple generalization runs.

To make it easier to identify directories that were created during
the generalization process, generated directories are now named
"teralizer_generated" instead of "teralizer" (which is already used
by the main teralizer project as well as the example projects).