Consolidate task scheduling in ProjectSetupTask
Most tasks are now scheduled in the ProjectSetupTask, which makes
it much easier to see at a glance whether the scheduled tasks are
correct, i.e., contain no duplicates and do not miss anything.

This is enabled, on the one hand, by the previously introduced
priority queue and, on the other hand, by the newly introduced
base-task vs. sub-task architecture.

Base tasks are project-level variants of the corresponding sub-tasks,
which makes it possible to schedule them in the ProjectSetupTask
before any Task- or GeneralizationRecords are known. The responsi-
bility of a base task is to schedule its corresponding sub-tasks, The
responsibility of a sub-task is to execute the actual task logic.