Fix detection of assertion errors in TestExecutionTask
When individual tests fail, the processing pipeline should
still continue processing of all other tests that did not fail.

To enable this, we need to detect whether a non-zero return value
of an `mvn test` command was caused by an assertion failure (=>
continue processing) or some other issue (=> abort processing).

This worked at some point in the past by checking whether the
message of corresponding ConsoleCommandExceptions contains the
String `AssertionFailedError`, which is thrown by JUnit 5 for
any failing assertions.

However, commit [1] broke this detection because since then, the full
failure text is not included in the ConsoleCommandException message
anymore, but is instead written to a file on disk.

To fix this, this commit now reads the failure text from disk and
then performs the usual assertion error detection on it. Furthermore,
we now also check JUnit 4 assertion failures (AssertionError) in
to JUnit 5 assertion failures (AssertionFailedError).