Improve handling of failed tests
Until now, failed tests always caused the processing pipeline to fail
during mutation testing. This was because PIT requires that all tests
that it processes during mutation testing are passing.

To avoid such failures, we are now only excluding test classes that
contain any failing tests from further processing.

Note that this means that we also exclude passing tests that are part
of the same class as some failing test. This is because PIT has no
method-level inclusion/exclusion settings, so we need to make do with
the class-level inclusions/exclusions.

---

Apart from the mutation testing issues, this commit also fixes the
issue that no coverage data was available if any test failed.

The cause of this issue was that JaCoCo's reporting task was never
executed when any test failed because the task was defined as a
Maven "goal" / Gradle "finalizedBy" of the test task. Thus, when the
test task failed (due to failing tests), Maven / Gradle also skipped
the execution of the JaCoCo reporting task.

To avoid this, we are now executing the test task and the JaCoCo
reporting task separately ourselves. This way, we can ensure that
the reporting task is executed even if some tests have failed.

---

Finally, we now no longer identify potential tests for generalization
by going through the test sources directory but instead collect them
from the test results that are reported by the test task.

This ensures that we are generalizing (or at least attempting to
generalize) exactly those tests that are actually executed by the
project. After all, some tests might deliberately be skipped or
otherwise excluded by the project configuration, so we should also
ignore them for generalization purposes.