Rename stages + classes + DB tables to contain used tool names
This is mostly a preparatory step for future changes that will add
PIT coverage reports to the DB. Since we already have JaCoCo coverage
reports stored there, we need some way to distinguish them.

As a side effect, it's now easier to see at a glance which processing
stages / classes / DB tables are related to which of the used tools,
which might have been nice to have anyway.

This commit is purely a refactoring / renaming - it contains no
functional changes that affect the behavior of the program.