Prevent TooManyFilterMissesExceptions from failing the pipeline
TooManyFilterMissesExceptions occur when jqwik is unable to generate
input values that satisfy the precondition / input specification.

While this is not ideal (the execution of the test just consumes time
for no benefit), it's also not a big enough issue that the pipeline
should fail as a result of it.

After all, the data that we want / need can still be collected even
if individual generalizations do not prove to be useful.