Use fixed seeds for jqwik properties
While use of fixed seeds is a good idea anyway for reproducibility
reason, we are actually REQUIRED to use fixed seeds.

This is because jqwik's property-base tests can fail with a
TooManyFilterMissesException when jqwik is unable to generate
input values that satisfy the precondition / input specification.

On its own, this would not be too bad, but PIT requires all tests
to pass for mutation testing purposes. To get around this, we filter
all failing tests and only apply PIT to passing ones.

However, the TooManyFilterMissesException might not occur when we
make our filtering decision, but then occur when running PIT, causing
PIT - and in turn, the overall pipeline - to fail. Use of a fixed seed
avoids this since TooManyFilterMissesException are then guaranteed not
to occur during the PIT run if they did not occur during filtering.