Exclude uncovered classes from mutation testing
PIT itself already implements a similar functionality, but we observed
that we get faster overall runtimes if we do the exclusion ourselves
without relying on PIT for it.

This is likely for (at least) the following reasons:

1. PIT checks coverage AFTER generating mutants (but before executing
   covering tests), so we save some time by avoiding the mutant
   generation overhead.

2. PIT generates data for all mutants, and we also mirror it in our
   database. Excluding uncovered classes before they get to PIT
   means PIT does not have to write any data for them, and we do
   not have to import the data to the DB.

Of course, this means that we do not get the "NO_COVERAGE" results
that PIT would provide when detecting that a mutant is not covered
by any test. However, such results are not very useful anyway since
we already know from JaCoCo that no mutants (whatever they may be)
will be covered by the original or generalized tests.