Add new task TestGenerationTask
The new task uses EvoSuite [1] to generate tests for a project.

By default, test generation is disabled but can be enabled on a
per-project basis by setting project.use_test_generation to true.

Because PIT does not play well with the isolation features of the
EvoSuite-generated tests, we have to do some post-processing on the
generated tests to remove these isolation features.

While this potentially reduces the robustness of the generated test
suite, this is fine for our purposes because the robustness of the
initial test suite does / should not affect generalization results.

On the plus side, this post-processing allows us to keep using PIT
for mutation testing, which is good because the planned evaluation
relies on mutation testing results to measure the effectiveness of
the test generalization approach. :)

[1]: https://www.evosuite.org/