Automatically create directories for EvoSuite-generated tests
We automatically copy the EvoSuite-generated tests from the data
directory (in which we "log" them) to the corresponding location
in the tests directory of the processed project (from which they are
deleted again at cleanup).

Thus, we need to ensure that the necessary directories that depict the
package hierarchy of the generated tests are available before copying.