Match test inclusions in pre-/post-generalization reports
The only difference between pre- and post-generalization reports
should be that pre-generalization reports do NOT consider generalized
tests but post-generalization ones DO.

Notably, this means that the in-/excluded ORIGINAL tests MUST NOT
differ across pre- and post-generalization reports. That way, any
differences that are observed are guaranteed to be caused by the
in-/exclusion of GENERALIZED tests (and not ORIGINAL ones).

Prior to this commit, this requirement was not fulfilled because pre-
generalization reports were collected BEFORE the JPF execution even
though JPF execution could lead to further exclusions of original
tests. Thus, the pre-generalization reports included more original
tests than the post-generalization ones.

With this commit, this mismatch is now fixed by collecting the pre-
generalization reports AFTER (instead of BEFORE) the JPF execution.