Add step/stage/projectId columns to some tables
The *.step / *.stage columns allow us to identify at which point
throughout the processing pipeline the corresponding reports were
collected, which was not possible to reliably identify before.

The assertion.project_id column is just for easier joining / grouping
because the project of an assertion could already be identified by
joining assertion -> test -> project.