Create generalizations in the same directories as the original tests
Until now, generalized classes were create in a sub-package of the
package that the original test classes are part of. This caused
compilation errors if (i) the test classes are in the same package
as the tested classes and (ii) the test classes use package-private
fields / methods / ... of the tested classes.

To avoid such compilation errors, we are now creating the generalized
classes in the same package as the original classes. This ensures
that the generalized classes can access the same (package-private)
fields / methods / ... as the original classes.

Similar reasoning holds for the driver classes, which might need to
access (package-private) test methods. Thus, driver classes are now
created in the same package as the corresponding test class as well.