Add variable-based constraints and equalities to IMPROVED variant
The initial version of the IMPROVED generalization variant only
used constant-based inequalities such as 'x < 10' and 'y >= 5' for
the generation of PIT arbitraries.

With this commit, the following are now also considered:
(i) variable-based constraints such as 'x <= y' or 'i > j',
(ii) equalities such as 'a = 15' or 'b = c'.

More complex (in-)equalities such as 'x < y + 1' or '(a * b) = c'
are still ignored, i.e., do not affect the arbitrary generation.