Fix Velocity macro name conflicts
Velocity macro definitions are global and cannot (easily) be redefined
after they have been set / loaded for the first time.

This caused the templates of the IMPROVED variant to use the macro
definitions of the NAIVE template because the processing steps for
the IMPROVED variant are executed after the NAIVE one.

To fix this, we are now using different names for the relevant
macro definitions of the IMPROVED vs. NAIVE variant.