Disable automated generation of jOOQ files
The issue with automated execution of the `generateJooq` task is that
it runs before any of the Teralizer code is executed.

If no database has been created yet, the result is that the existing
jOOQ files (which are included in the repository) are deleted, causing
the following project build + execution to fail.

Because we don't want to include the DB in the repository, but would
rather create it automatically through a Teralizer task that uses the
included DDL scripts, running jOOQ generation automatically is,
therefore, not an option.