Add wrapper methods that uniquely mark the symbolic.method target
To avoid issues with tests that call a "tested method" multiple times,
we now wrap each tested method call in a custom marker method that can
be used for SPF's "symbolic.method" setting.

For example:

```java
@Test
void test() {
  assertEquals(..., Math.abs(...));
  assertEquals(..., Math.abs(...));
}
```

Is now transformed to something like the following for the first
JPF run that's supposed to extract the specification for the first
Math.abs call:

```java
@Test
void test() {
  assertEquals(..., abs_marker(...));
  assertEquals(..., Math.abs(...));
}

... abs_marker(...) {
  return Math.abs(...);
}
```

This ensures that only constraints from the first Math.abs call
are added to the path condition during this JPF run.