Fix marking of tests without assertions as excluded
Until now, tests without assertions were (supposed to be) marked as
excluded by the AssertionCountFilter / TestAnalysisTask.

This broke with 6f7ab3e2dc37ba298f6992e1657b3dabf39a521e because since
that commit, the TestFilterTask is only applied to tests that contain
at least one assertion (we are actually filtering assertions rather
than tests, because assertions are now the generalization target).

To restore this functionality, the TestAnalysisTask now throws an
exception if any of the processed tests do not contain any assertions.