Exclude anonymous classes from mutation testing
We're using JaCoCo reports to figure out which classes are targeted
by the included tests. If anonymous classes are targeted, their names
are represented like the following in these reports:

com.example.MyClass.new MyInterface() {...}

That is, anonymous classes are shown as "new <InterfaceName> {...}".
When passing such anonymous class names to the `targetClases` argument
of the `mvn pitest:mutationCoverage` command, PIT fails with an
'Illegal repetition' error (probably due to the "...").

To avoid such errors, we are now excluding all class names that
contain the string "...".