Add tested method identification for assertThrows
assertThrows contain the tested method in an executable passed on as the second method parameter.
There are several options to prepare this argument:
+ lambda expression () -> method() or () -> {method();}
+ method reference this::method
+ anonymous inner class
+ by referencing a variable

If an object is needed for the call it can be constructed in the executable which would lead to multiple invocations. Thus, the tested method is assumed to be the last method invocation in the executable.