Remove inclusion of asserThrows with method reference
Method references like this::someMethod can be used to specify an executable in an assertThrows call.
Previously the referenced method was analyzed to find the tested method within.
However, it is common to supply the tested method itself using this technique.
Because we cannot check whether the tested method is specified by the reference or contained within the referenced method, we should not include the assertion in further analysis.