Fix NullPointerException in TestGeneralizationTask
When a test is generalized where the tested method has method parameters that cannot be generalized e.g. arrays,
the generation of the TestParameters class was not successful. Now, the parameters are filtered before creating the TestParameters class such that only generalizable parameters are used.

Fixes issue: https://github.com/glockyco/test-generalization/issues/121