Filter EvoSuite tests targetting "Undeclared Exception"s
EvoSuite tests with "Undeclared Exception!" comments are intended
to reveal thrown Exceptions that are not shown in the throws clause
of a method signature. As such, these tests are intended to FAIL.

Since PIT only supports mutation testing for classes that have no
failing tests, we now remove these tests during postprocessing of
the EvoSuite-generated test suite.

This way, we can at least keep the remaining non-failing tests that
were generated by EvoSuite (keeping the failing tests in the test
suite but excluding them during mutation testing is not an option
because PIT only supports class-level exclusions).