Remove "()" from all `testMethodName`s
In Maven projects, testCaseReport.getName() always returned the method
name WITHOUT a trailing "()" already. However, in Gradle projects,
the method names were returned WITH  a trailing "()".

To have a normalized naming scheme to work with throughout the
processing pipeline, this commit now removes any "()" that is present
in the method name.

The variant without "()" is preferred because, e.g., Spoon's
`CtClass.getMethod(methodName)` method expects the method name to
be the one without "()".