Set Java compile version in Gradle example projects
For some reason Gradle did not use the Java compiler set using jenv
to compile the example project.

This resulted in class files compiled with a different java version.

To ensure that Java 8 is used to compile the project files, the
language version is now set in the build.gradle file, similarly
to the compiler configuration in the Maven POM files.