Add baseline generalization variant
This variant uses an arbitrary supplier that just returns the
concrete input values used in the original tests, and does
not modify the expected values of assertions.

It is intended to be used as (i) a "baseline" for  generalization
failure analysis and (ii) an intermediate reference value for runtime
comparisons of original vs. (more sophisticated) generalized tests.