Add per-variant configuration support
This enables the same generalization "algorithm" (BASELINE, NAIVE,
IMPROVED) to be executed multiple times with different configuration
settings in a single run of the tool.

Thus, each "variant" can now be characterized not only by the
algorithm that it uses, but also by other configuration settings
(which have yet to be added in future commits).