Fix baseline generalization creating incompilable code
The issue was that the BaselineParametersSupplierFactory used the
filtered parameters (which only contain parameters that DO match the
supported types) with the unfiltered arguments (which also contain
arguments that do NOT match the supported types).

As a result, there was often a mismatch between (types of) parameters
and arguments in the code, causing compilation to fail.

This commit fixes such mismatches by using the arguments map (rather
than list) to select the correct argument for a given parameter.