Use the same targetClasses in all non-original PIT runs
The intention of this is to make sure that results from
COLLECT_PIT_DATA_INITIAL and COLLECT_PIT_DATA_GENERALIZED are
directly comparable by ensuring they have the same number of mutants.

This broke with the introduction of the COLLECT_PIT_DATA_ORIGINAL
in commit 8cd3f17056afedc05c1a445629d0815866dcdae6.

Since then, target classes from ORIGINAL and INITIAL were combined,
basically resulting in INITIAL using the same target classes as
ORIGINAL (because INITIAL is always a subset of ORIGINAL).

To fix this, we're now distinguishing based on not only the variant
(which is null for both ORIGINAL as well as INITIAL) but also based
on the stage (which is different for these two).

Additionally, we explicitly use the target classes from INITIAL
for the GENERALIZED variants. While this might not be absolutely
necessary because INITIAL and GENERALIZED *should* have the same
target classes anyway, this makes the intention clearer and is
less likely to break again in the future.