Add `filter_result` table to the DB
Until now, information about all filter results associated with a
test, assertion, or generalization was stored a single text string
in the `task.info` column.

To enable us to more easily process / analyze filtering results,
this commit moves filter result data from the `task.info` column
to a new `filter_result` table.

The overall information that is stored about a filtering result
is still the same, so no data is added or removed.