Remove "DB file exists" checks from ProcessingPipeline
These checks were originally added to avoid SQLite DB files from
being (re-)created if they don't exist. The two most common reasons
for this back then were that (i) no DB had yet been created at all
and (ii) the DB was removed by a CleanupTask.

While this idea of preserving a clean state might have some merit,
it was only ever enforced in the ProcessingPipeline class, but not
anywhere else in the code base. Consequently, the DB file would still
be recreated by jOOQ when storing other (i.e., non-task) records.

Additionally, with the recent migration from SQLite to PostgreSQL,
there is even less of a reason to attempt to keep a clean state
based on a "DB file exists" check, since the existence of a DB
"file" (it is now a directory) has no reasonable connection to
whether we are in a cleaned-up or non-cleanup state.

Finally, the CleanupTask hasn't removed the DB for a long time even
while we were still using SQLite because we want to preserve historic
data even past a cleanup. Thus, the notion of "cleaned-up" == "no DB
file exists" has not been useful even before the Postgres migration.