Fix `mutation_id` calculation
`mutated_package` needs to be considered to distinguish classes
with the same name but different packages. This broke when splitting
the class and package name in [1]. Before that, `mutated_class`
contained the fully qualified class name.

`mutated_method` is not really necessary to consider, because the
`line_number` refers to the location in the class. However, including
the method as well makes the ordering more robust against regressions.

`method_description` should have been considered all along, but was
simply missed when the view was originally implemented.

[1]: 33b91c48822d44596e8b45fa0fb37155543a0ad7