Fix incompilable code for methods with generic targetType
Instead of producing incompilable code when generating the wrapper
method for JPF/SPF execution, we are now instead marking such cases
as excluded by throwing a corresponding exception.

As a result, we can still continue processing of other methods in the
project that were successfully instrumented, rather than having to
stop the whole processing pipeline due to incompilable code.