Improve tested method detection for re-assigned assertion params
This improves detection of the tested method for cases such as:

public void testReadLeftAndWhiteSpace_Various() {
  int exp = 0;
  int act = ParserUtil.readLeftWhiteSpace(sb2, 0, sb2.length()-1);
  assertEquals(exp, act);

  exp = sb2.length()-1;
  act = ParserUtil.readRightWhiteSpace(sb2, 0, sb2.length()-1);
  assertEquals(exp, act);
}

The old implementation always identified the first call to
ParserUtil.readLeftWhiteSpace as the tested method call for
BOTH assertEquals calls.

The new implementation instead identifies the second call to
ParserUtil.readLeftWhiteSpace as the tested method call of the
second assertEquals call.